remove New*** from variable name.
