+++ executor/aggregate_xapi.go
@@ -114,31 +114,43 @@ func (n *finalAggregater) updateSum(val types.Datum, count uint64) error {
 	return nil
 }
 
+// The argument if the name of a aggregate function.
+// This function will check if the aggregate function need count in partial result.
+func needCount(name string) bool {
+	return name == ast.AggFuncCount || name == ast.AggFuncAvg
+}
+
+// The argument if the name of a aggregate function.
+// This function will check if the aggregate function need value in partial result.
+func needValue(name string) bool {
+	return name == ast.AggFuncSum || name == ast.AggFuncAvg || name == ast.AggFuncFirstRow ||
+		name == ast.AggFuncMax || name == ast.AggFuncMin || name == ast.AggFuncGroupConcat
+}
+
 // XAggregateExec deals with all the aggregate functions.
 // It is built from Aggregate Plan. When Next() is called, it reads all the data from Src and updates all the items in AggFuncs.
-// TODO: Support having.
 type XAggregateExec struct {
 	Src               Executor
-	ResultFields      []*ast.ResultField
 	AggFields         []*types.FieldType
 	executed          bool
 	ctx               context.Context
-	AggFuncs          []*ast.AggregateFuncExpr
+	AggFuncs          []expression.AggregationFunction
 	aggregaters       []*finalAggregater
 	groupMap          map[string]bool
 	groups            [][]byte
 	hasGroupBy        bool
 	currentGroupIndex int
+	schema            expression.Schema
 }
 
 // Schema implements Executor Schema interface.
 func (e *XAggregateExec) Schema() expression.Schema {
-	return nil
+	return e.schema
 }
 
 // Fields implements Executor Fields interface.
 func (e *XAggregateExec) Fields() []*ast.ResultField {
-	return e.ResultFields
+	return nil
 }
 
 // Next implements Executor Next interface.
@@ -150,7 +162,7 @@ func (e *XAggregateExec) Next() (*Row, error) {
 		e.aggregaters = make([]*finalAggregater, len(e.AggFuncs))
 		for i, af := range e.AggFuncs {
 			agg := &finalAggregater{
-				name:               strings.ToLower(af.F),
+				name:               strings.ToLower(af.GetName()),
 				contextPerGroupMap: make(map[string](*ast.AggEvaluateContext)),
 			}
 			e.aggregaters[i] = agg
@@ -180,12 +192,13 @@ func (e *XAggregateExec) Next() (*Row, error) {
 	if e.currentGroupIndex >= len(e.groups) {
 		return nil, nil
 	}
+	retRow := &Row{Data: make([]types.Datum, 0, len(e.AggFuncs))}
 	groupKey := e.groups[e.currentGroupIndex]
 	for _, af := range e.AggFuncs {
-		af.CurrentGroup = groupKey
+		retRow.Data = append(retRow.Data, af.GetGroupResult([]byte(groupKey)))
 	}
 	e.currentGroupIndex++
-	return &Row{}, nil
+	return retRow, nil
 }
 
 // Fetch a single row from src and update each aggregate function.
@@ -239,153 +252,6 @@ func (e *XAggregateExec) innerNext() (bool, error) {
 
 // Close implements Executor Close interface.
 func (e *XAggregateExec) Close() error {
-	for _, af := range e.AggFuncs {
-		af.Clear()
-	}
-	if e.Src != nil {
-		return e.Src.Close()
-	}
-	return nil
-}
-
-// The argument if the name of a aggregate function.
-// This function will check if the aggregate function need count in partial result.
-func needCount(name string) bool {
-	return name == ast.AggFuncCount || name == ast.AggFuncAvg
-}
-
-// The argument if the name of a aggregate function.
-// This function will check if the aggregate function need value in partial result.
-func needValue(name string) bool {
-	return name == ast.AggFuncSum || name == ast.AggFuncAvg || name == ast.AggFuncFirstRow ||
-		name == ast.AggFuncMax || name == ast.AggFuncMin || name == ast.AggFuncGroupConcat
-}
-
-// NewXAggregateExec deals with all the aggregate functions.
-// It is built from Aggregate Plan. When Next() is called, it reads all the data from Src and updates all the items in AggFuncs.
-type NewXAggregateExec struct {
-	Src               Executor
-	AggFields         []*types.FieldType
-	executed          bool
-	ctx               context.Context
-	AggFuncs          []expression.AggregationFunction
-	aggregaters       []*finalAggregater
-	groupMap          map[string]bool
-	groups            [][]byte
-	hasGroupBy        bool
-	currentGroupIndex int
-	schema            expression.Schema
-}
-
-// Schema implements Executor Schema interface.
-func (e *NewXAggregateExec) Schema() expression.Schema {
-	return e.schema
-}
-
-// Fields implements Executor Fields interface.
-func (e *NewXAggregateExec) Fields() []*ast.ResultField {
-	return nil
-}
-
-// Next implements Executor Next interface.
-func (e *NewXAggregateExec) Next() (*Row, error) {
-	// In this stage we consider all data from src as a single group.
-	if !e.executed {
-		e.groupMap = make(map[string]bool)
-		e.groups = [][]byte{}
-		e.aggregaters = make([]*finalAggregater, len(e.AggFuncs))
-		for i, af := range e.AggFuncs {
-			agg := &finalAggregater{
-				name:               strings.ToLower(af.GetName()),
-				contextPerGroupMap: make(map[string](*ast.AggEvaluateContext)),
-			}
-			e.aggregaters[i] = agg
-		}
-		for {
-			hasMore, err := e.innerNext()
-			if err != nil {
-				return nil, errors.Trace(err)
-			}
-			if !hasMore {
-				break
-			}
-		}
-		e.executed = true
-		if len(e.groups) == 0 && !e.hasGroupBy {
-			// If no groupby and no data, we should add an empty group.
-			// For example:
-			// "select count(c) from t;" should return one row [0]
-			// "select count(c) from t group by c1;" should return empty result set.
-			e.groups = append(e.groups, singleGroup)
-		}
-		// Set AggregateFuncExprs' context.
-		for i, agg := range e.aggregaters {
-			e.AggFuncs[i].SetContext(agg.contextPerGroupMap)
-		}
-	}
-	if e.currentGroupIndex >= len(e.groups) {
-		return nil, nil
-	}
-	retRow := &Row{Data: make([]types.Datum, 0, len(e.AggFuncs))}
-	groupKey := e.groups[e.currentGroupIndex]
-	for _, af := range e.AggFuncs {
-		retRow.Data = append(retRow.Data, af.GetGroupResult([]byte(groupKey)))
-	}
-	e.currentGroupIndex++
-	return retRow, nil
-}
-
-// Fetch a single row from src and update each aggregate function.
-// If the first return value is false, it means there is no more data from src.
-func (e *NewXAggregateExec) innerNext() (bool, error) {
-	var row *Row
-	if e.Src != nil {
-		var err error
-		row, err = e.Src.Next()
-		if err != nil {
-			return false, errors.Trace(err)
-		}
-		if row == nil {
-			return false, nil
-		}
-	} else {
-		// If Src is nil, only one row should be returned.
-		if e.executed {
-			return false, nil
-		}
-	}
-	e.executed = true
-	// cursor is used to traverse the row.
-	var cursor int
-	// The first column is groupkey.
-	groupKey := row.Data[cursor].GetBytes()
-	if _, ok := e.groupMap[string(groupKey)]; !ok {
-		e.groupMap[string(groupKey)] = true
-		e.groups = append(e.groups, groupKey)
-	}
-	cursor++
-	// The rest columns are partial result for aggregate function.
-	for _, agg := range e.aggregaters {
-		var count uint64
-		var value types.Datum
-		if needCount(agg.name) {
-			// count partial result field
-			count = row.Data[cursor].GetUint64()
-			cursor++
-		}
-		if needValue(agg.name) {
-			// value partial result field
-			value = row.Data[cursor]
-			cursor++
-		}
-		agg.currentGroup = groupKey
-		agg.update(count, value)
-	}
-	return true, nil
-}
-
-// Close implements Executor Close interface.
-func (e *NewXAggregateExec) Close() error {
 	e.executed = false
 	e.groups = nil
 	e.currentGroupIndex = 0
